KafkaSchemaBasedSerializationSchemaFactory simpler
