+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/SchemaBasedSerdeProvider.scala
@@ -1,13 +1,13 @@
 package pl.touk.nussknacker.engine.schemedkafka.schemaregistry
 
-import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.UniversalToJsonFormatterFactory
-import pl.touk.nussknacker.engine.schemedkafka.serialization.{
-  KafkaSchemaBasedKeyValueDeserializationSchemaFactory,
-  KafkaSchemaBasedSerializationSchemaFactory
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.{
+  UniversalKafkaSerializationSchemaFactory,
+  UniversalToJsonFormatterFactory
 }
+import pl.touk.nussknacker.engine.schemedkafka.serialization.KafkaSchemaBasedKeyValueDeserializationSchemaFactory
 
 case class SchemaBasedSerdeProvider(
-    serializationSchemaFactory: KafkaSchemaBasedSerializationSchemaFactory,
+    serializationSchemaFactory: UniversalKafkaSerializationSchemaFactory,
     deserializationSchemaFactory: KafkaSchemaBasedKeyValueDeserializationSchemaFactory,
     recordFormatterFactory: UniversalToJsonFormatterFactory,
 )
