+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/schemaregistry/universal/UniversalKafkaSerializationSchemaFactory.scala
@@ -1,0 +1,57 @@
+package pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal
+
+import io.confluent.kafka.schemaregistry.ParsedSchema
+import org.apache.kafka.clients.producer.ProducerRecord
+import org.apache.kafka.common.header.internals.RecordHeaders
+import org.apache.kafka.common.serialization.Serializer
+import pl.touk.nussknacker.engine.api.process.TopicName
+import pl.touk.nussknacker.engine.kafka.{serialization, KafkaConfig}
+import pl.touk.nussknacker.engine.kafka.serialization.{CharSequenceSerializer, KafkaProducerHelper}
+import pl.touk.nussknacker.engine.schemedkafka.RuntimeSchemaData
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.SchemaRegistryClientFactory
+import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.confluent.schemaid.SchemaIdFromNuHeadersPotentiallyShiftingConfluentPayload.ValueSchemaIdHeaderName
+import pl.touk.nussknacker.engine.util.KeyedValue
+
+import java.lang
+import java.nio.charset.StandardCharsets
+
+class UniversalKafkaSerializationSchemaFactory(
+    protected val schemaRegistryClientFactory: SchemaRegistryClientFactory,
+) extends Serializable {
+
+  def create(
+      topic: TopicName.ForSink,
+      valueSchemaData: RuntimeSchemaData[ParsedSchema],
+      kafkaConfig: KafkaConfig
+  ): serialization.KafkaSerializationSchema[KeyedValue[AnyRef, AnyRef]] = {
+    new serialization.KafkaSerializationSchema[KeyedValue[AnyRef, AnyRef]] {
+      private lazy val keySerializer   = new CharSequenceSerializer
+      private lazy val valueSerializer = createValueSerializer(valueSchemaData, kafkaConfig)
+
+      override def serialize(
+          element: KeyedValue[AnyRef, AnyRef],
+          timestamp: lang.Long
+      ): ProducerRecord[Array[Byte], Array[Byte]] = {
+        // we have to create headers for each record because it can be enriched (mutated) by serializers
+        val headers = new RecordHeaders()
+        val key     = keySerializer.serialize(topic.name, headers, element.key)
+        valueSchemaData.schemaIdOpt
+          .foreach(id => headers.add(ValueSchemaIdHeaderName, id.toString.getBytes(StandardCharsets.UTF_8)))
+        val value = valueSerializer.serialize(topic.name, headers, element.value)
+        KafkaProducerHelper.createRecord(topic.name, key, value, timestamp, headers)
+      }
+    }
+  }
+
+  private def createValueSerializer(
+      schemaData: RuntimeSchemaData[ParsedSchema],
+      kafkaConfig: KafkaConfig
+  ): Serializer[Any] = {
+    val schemaRegistryClient = schemaRegistryClientFactory.create(kafkaConfig)
+    val schema               = schemaData.schema
+    UniversalSchemaSupportDispatcher(kafkaConfig)
+      .forSchemaType(schema.schemaType())
+      .serializer(Some(schema), schemaRegistryClient, isKey = false)
+  }
+
+}
